Final refactor to load from binary
