+++ LeapOfFaith/config.json
@@ -1,5 +1,5 @@
 {
   "width": 1366,
   "height": 768,
-  "model": "models/uh60.obj"
+  "model": "models/uh60.bin"
 }
