+++ LeapOfFaithLib/LeapOfFaithLib.vcxproj
@@ -108,6 +108,8 @@
     <ClInclude Include="src\Engine.h" />
     <ClInclude Include="src\Mesh.h" />
     <ClInclude Include="src\MeshModel.h" />
+    <ClInclude Include="src\MeshReader.h" />
+    <ClInclude Include="src\TextureUtils.h" />
     <ClInclude Include="src\Utilities.h" />
     <ClInclude Include="src\VulkanRenderer.h" />
   </ItemGroup>
@@ -115,7 +117,9 @@
     <ClCompile Include="src\Engine.cpp" />
     <ClCompile Include="src\Mesh.cpp" />
     <ClCompile Include="src\MeshModel.cpp" />
+    <ClCompile Include="src\MeshReader.cpp" />
     <ClCompile Include="src\VulkanRenderer.cpp" />
+    <ClCompile Include="src\TextureUtils.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="src\shaders\compile_shaders.bat" />
