+++ LeapOfFaithLib/src/Mesh.h
@@ -14,7 +14,6 @@ struct Model {
 class Mesh
 {
 public:
-	Mesh();
 	Mesh(VkPhysicalDevice newPhysicalDevice, VkDevice newDevice, 
 		VkQueue transferQueue, VkCommandPool transferCommandPool, 
 		std::vector<Vertex>* vertices, std::vector<uint32_t>* indices,
