+++ LeapOfFaithLib/src/MeshModel.cpp
@@ -1,11 +1,6 @@
-#define STB_IMAGE_IMPLEMENTATION
-#include <stb_image.h>
-
-#include <assimp/Importer.hpp>
-#include <assimp/scene.h>
-#include <assimp/postprocess.h>
-
 #include "MeshModel.h"
+#include "MeshReader.h"
+#include "TextureUtils.h"
 
 MeshModel::MeshModel()
 {
@@ -16,134 +11,12 @@ void MeshModel::LoadFile(const std::string& modelFile, VkPhysicalDevice& physica
 	std::vector<VkImage>& textureImages, std::vector<VkDeviceMemory>& textureImageMemory, std::vector<VkImageView>& textureImageViews,
 	VkDescriptorPool& samplerDescriptorPool, VkDescriptorSetLayout& samplerSetLayout, VkSampler& textureSampler, std::vector<VkDescriptorSet>& samplerDescriptorSets)
 {
-	// Import model "scene"
-	Assimp::Importer importer;
-	const aiScene* scene = importer.ReadFile(modelFile, aiProcess_Triangulate | aiProcess_FlipUVs | aiProcess_JoinIdenticalVertices);
-	if (!scene)
-	{
-		throw std::runtime_error("Failed to load model! (" + modelFile + ")");
-	}
-
-	// Get vector of all materials with 1:1 ID placement
-	std::vector<std::string> textureNames = MeshModel::LoadMaterials(scene);
-
-	// Conversion from the materials list IDs to our Descriptor Array IDs
-	std::vector<int> matToTex(textureNames.size());
-
-	// Loop over textureNames and create textures for them
-	for (size_t i = 0; i < textureNames.size(); i++)
-	{
-		// If material had no texture, set '0' to indicate no texture, texture 0 will be reserved for a default texture
-		if (textureNames[i].empty())
-		{
-			matToTex[i] = 0;
-		}
-		else
-		{
-			// Otherwise, create texture and set value to index of new texture
-			matToTex[i] = createTexture(textureNames[i], physicalDevice, logicalDevice, graphicsQueue, graphicsCommandPool, 
-				textureImages,textureImageMemory, textureImageViews, samplerDescriptorPool, samplerSetLayout, textureSampler, samplerDescriptorSets);
-		}
-	}
-
 	// Load in all our meshes
-	std::vector<Mesh> modelMeshes = MeshModel::LoadNode(physicalDevice, logicalDevice, graphicsQueue, graphicsCommandPool,
-		scene->mRootNode, scene, matToTex);
+	std::vector<Mesh> meshList;
+	MeshReader::loadFromBinary(modelFile, meshList, physicalDevice, logicalDevice, graphicsQueue, graphicsCommandPool, 
+		textureImages, textureImageMemory, textureImageViews, samplerDescriptorPool, samplerSetLayout, textureSampler, samplerDescriptorSets);
 
-	this->meshList = modelMeshes;
-}
-
-int MeshModel::createTexture(const std::string& fileName, VkPhysicalDevice& physicalDevice, VkDevice& logicalDevice, VkQueue& graphicsQueue, VkCommandPool& graphicsCommandPool,
-	std::vector<VkImage>& textureImages, std::vector<VkDeviceMemory>& textureImageMemory, std::vector<VkImageView>& textureImageViews, 
-	VkDescriptorPool& samplerDescriptorPool, VkDescriptorSetLayout& samplerSetLayout, VkSampler& textureSampler, std::vector<VkDescriptorSet>& samplerDescriptorSets)
-{
-	// Create Texture Image and get its location in array
-	int textureImageLoc = MeshModel::createTextureImage(fileName, physicalDevice, logicalDevice, graphicsQueue, graphicsCommandPool, textureImages, textureImageMemory);
-
-	// Create image view and add to list
-	VkImageView imageView = createImageView(textureImages[textureImageLoc], VK_FORMAT_R8G8B8A8_UNORM, VK_IMAGE_ASPECT_COLOR_BIT, logicalDevice);
-	textureImageViews.push_back(imageView);
-
-	// Create texture descriptor
-	int descriptorLoc = createTextureDescriptor(imageView, samplerDescriptorPool, logicalDevice, samplerSetLayout, textureSampler, samplerDescriptorSets);
-
-	// Return location of set with texture
-	return descriptorLoc;
-}
-
-VkImageView MeshModel::createImageView(VkImage image, VkFormat format, VkImageAspectFlags aspectFlags, VkDevice& logicalDevice)
-{
-	VkImageViewCreateInfo viewCreateInfo = {};
-	viewCreateInfo.sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO;
-	viewCreateInfo.image = image; // Image to create view for
-	viewCreateInfo.viewType = VK_IMAGE_VIEW_TYPE_2D; // Type of image (1D, 2d, 3D, Cube, etc)
-	viewCreateInfo.format = format; // Format of image data
-	viewCreateInfo.components.r = VK_COMPONENT_SWIZZLE_IDENTITY; // Allows remmapping of rgba components to other rgba values
-	viewCreateInfo.components.g = VK_COMPONENT_SWIZZLE_IDENTITY;
-	viewCreateInfo.components.b = VK_COMPONENT_SWIZZLE_IDENTITY;
-	viewCreateInfo.components.a = VK_COMPONENT_SWIZZLE_IDENTITY;
-
-	// Subresources allow the view to view only a part of an image
-	viewCreateInfo.subresourceRange.aspectMask = aspectFlags; // which aspect of the image to view, (COLOR_BIT etc)
-	viewCreateInfo.subresourceRange.baseMipLevel = 0; // start mipmap level to view from
-	viewCreateInfo.subresourceRange.levelCount = 1; // Number of mipmap levels to view
-	viewCreateInfo.subresourceRange.baseArrayLayer = 0; // Start array level to view from
-	viewCreateInfo.subresourceRange.levelCount = 1; // Number of array levels to view
-
-	// Create image view and return it
-	VkImageView imageView;
-	VkResult result = vkCreateImageView(logicalDevice, &viewCreateInfo, nullptr, &imageView);
-	if (result != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to create an Image View!");
-	}
-
-	return imageView;
-}
-
-int MeshModel::createTextureDescriptor(VkImageView textureImage, VkDescriptorPool& samplerDescriptorPool, VkDevice& logicalDevice, 
-	VkDescriptorSetLayout& samplerSetLayout, VkSampler& textureSampler, std::vector<VkDescriptorSet>& samplerDescriptorSets)
-{
-	VkDescriptorSet descriptorSet;
-
-	// Descriptor set allocation info
-	VkDescriptorSetAllocateInfo setAllocInfo = {};
-	setAllocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-	setAllocInfo.descriptorPool = samplerDescriptorPool;
-	setAllocInfo.descriptorSetCount = 1;
-	setAllocInfo.pSetLayouts = &samplerSetLayout;
-
-	// Allocate Descriptor Sets
-	VkResult result = vkAllocateDescriptorSets(logicalDevice, &setAllocInfo, &descriptorSet);
-	if (result != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to allocate Texture Descriptor Sets!");
-	}
-
-	// Texture Image info
-	VkDescriptorImageInfo imageInfo = {};
-	imageInfo.imageLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL; //Image layout when in use
-	imageInfo.imageView = textureImage; // Image to bind to set
-	imageInfo.sampler = textureSampler; // Sampler to use for set
-
-	// Descriptor write info
-	VkWriteDescriptorSet descriptorWrite = {};
-	descriptorWrite.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-	descriptorWrite.dstSet = descriptorSet;
-	descriptorWrite.dstBinding = 0;
-	descriptorWrite.dstArrayElement = 0;
-	descriptorWrite.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-	descriptorWrite.descriptorCount = 1;
-	descriptorWrite.pImageInfo = &imageInfo;
-
-	// Update new descriptor set
-	vkUpdateDescriptorSets(logicalDevice, 1, &descriptorWrite, 0, nullptr);
-
-	// Add descriptor set to list
-	samplerDescriptorSets.push_back(descriptorSet);
-
-	// Return descriptor set location
-	return samplerDescriptorSets.size() - 1;
+	this->meshList = meshList;
 }
 
 Mesh* MeshModel::getMesh(size_t index)
@@ -167,232 +40,4 @@ void MeshModel::destroyMeshModel()
 	{
 		mesh.destroyBuffers();
 	}
-}
-
-std::vector<std::string> MeshModel::LoadMaterials(const aiScene* scene)
-{
-	// Create 1:1 sized list of textures
-	std::vector<std::string> textureList(scene->mNumMaterials);
-
-	// Go through each material and copy its texture file name (if it exists)
-	for (size_t i = 0; i < scene->mNumMaterials; i++)
-	{
-		// Get the material
-		aiMaterial* material = scene->mMaterials[i];
-
-		// Initialize the texture to empty string (will be replaced if texture exists)
-		textureList[i] = "";
-
-		// Check for a Diffuse Texture (standard details texture)
-		if (material->GetTextureCount(aiTextureType_DIFFUSE))
-		{
-			// Get path of the texture file
-			aiString path;
-			if (material->GetTexture(aiTextureType_DIFFUSE, 0, &path) == AI_SUCCESS)
-			{
-				// Cut off any directory information already present
-				int idx = std::string(path.data).rfind("\\");
-				std::string fileName = std::string(path.data).substr(idx + 1);
-
-				textureList[i] = fileName;
-			}
-		}
-	}
-
-	return textureList;
-}
-
-std::vector<Mesh> MeshModel::LoadNode(VkPhysicalDevice newPhysicalDevice, VkDevice newDevice, VkQueue transferQueue, VkCommandPool transferCommandPool, aiNode* node, const aiScene* scene, std::vector<int> matToTex)
-{
-	std::vector<Mesh> meshList;
-
-	// Go through each mesh at this node and create it, then add it to our meshList
-	for (size_t i = 0; i < node->mNumMeshes; i++)
-	{
-		meshList.push_back(
-			LoadMesh(newPhysicalDevice, newDevice, transferQueue, transferCommandPool, scene->mMeshes[node->mMeshes[i]], scene, matToTex)
-		);
-	}
-
-	// Go through each node attached to this node and load it, then append their meshes to this node's mesh list
-	for (size_t i = 0; i < node->mNumChildren; i++)
-	{
-		std::vector<Mesh> newList = LoadNode(newPhysicalDevice, newDevice, transferQueue, transferCommandPool, node->mChildren[i], scene, matToTex);
-		meshList.insert(meshList.end(), newList.begin(), newList.end());
-	}
-
-	return meshList;
-}
-
-Mesh MeshModel::LoadMesh(VkPhysicalDevice newPhysicalDevice, VkDevice newDevice, VkQueue transferQueue, VkCommandPool transferCommandPool, aiMesh* mesh, const aiScene* scene, std::vector<int> matToTex)
-{
-	std::vector<Vertex> vertices;
-	std::vector<uint32_t> indices;
-
-	// Resize vertex list to hold all vertices for mesh
-	vertices.resize(mesh->mNumVertices);
-
-	// Go through each vertex and copy it across to our vertices
-	for (size_t i = 0; i < mesh->mNumVertices; i++)
-	{
-		// Set position
-		vertices[i].pos = { mesh->mVertices[i].x, mesh->mVertices[i].y, mesh->mVertices[i].z };
-
-		// Set tex coords (if they exist)
-		if (mesh->mTextureCoords[0])
-		{
-			vertices[i].tex = { mesh->mTextureCoords[0][i].x, mesh->mTextureCoords[0][i].y };
-		}
-		else
-		{
-			vertices[i].tex = { 0.0f, 0.0f };
-		}
-
-		// Set color (just use white for now)
-		vertices[i].col = { 1.0f, 1.0f, 1.0f };
-	}
-
-	// Iterate over indices through faces and copy across
-	for (size_t i = 0; i < mesh->mNumFaces; i++)
-	{
-		// Get a face
-		aiFace face = mesh->mFaces[i];
-
-		// Go through faces indices and add to list
-		for (size_t j = 0; j < face.mNumIndices; j++)
-		{
-			indices.push_back(face.mIndices[j]);
-		}
-	}
-
-	// Create new mesh with details and return it
-	Mesh newMesh = Mesh(newPhysicalDevice, newDevice, transferQueue, transferCommandPool, &vertices, &indices, matToTex[mesh->mMaterialIndex]);
-
-	return newMesh;
-}
-
-stbi_uc* MeshModel::loadTextureFile(const std::string& fileName, int* width, int* height, VkDeviceSize* imageSize)
-{
-	// Number of channels image uses
-	int channels;
-
-	// Load pixel data for image
-	std::string fileLoc = "textures/" + fileName;
-	stbi_uc* image = stbi_load(fileLoc.c_str(), width, height, &channels, STBI_rgb_alpha);
-
-	if (!image)
-	{
-		throw std::runtime_error("Failed to load a Texture file! (" + fileName + ")");
-	}
-
-	// Calculate image size using given and known data
-	*imageSize = (*width) * (*height) * 4;
-
-	return image;
-}
-
-int MeshModel::createTextureImage(const std::string& fileName, VkPhysicalDevice& physicalDevice, VkDevice& logicalDevice, 
-	VkQueue& graphicsQueue, VkCommandPool& graphicsCommandPool, std::vector<VkImage>& textureImages, std::vector<VkDeviceMemory>& textureImageMemory)
-{
-	// Load image file
-	int width, height;
-	VkDeviceSize imageSize;
-	stbi_uc* imageData = loadTextureFile(fileName, &width, &height, &imageSize);
-
-	// Create staging buffer to hold loaded data, ready to copy to device
-	VkBuffer imageStagingBuffer;
-	VkDeviceMemory imageStagingBufferMemory;
-	createBuffer(physicalDevice, logicalDevice, imageSize, VK_BUFFER_USAGE_TRANSFER_SRC_BIT,
-		VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT,
-		&imageStagingBuffer, &imageStagingBufferMemory);
-
-	// Copy image to staging buffer
-	void* data;
-	vkMapMemory(logicalDevice, imageStagingBufferMemory, 0, imageSize, 0, &data);
-	memcpy(data, imageData, static_cast<size_t>(imageSize));
-	vkUnmapMemory(logicalDevice, imageStagingBufferMemory);
-
-	// Free original image data
-	stbi_image_free(imageData);
-
-	// Create image to hold final texture
-	VkImage texImage;
-	VkDeviceMemory texImageMemory;
-	texImage = createImage(width, height, VK_FORMAT_R8G8B8A8_UNORM, VK_IMAGE_TILING_OPTIMAL,
-		VK_IMAGE_USAGE_TRANSFER_DST_BIT | VK_IMAGE_USAGE_SAMPLED_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT, &texImageMemory, physicalDevice, logicalDevice);
-
-	// Copy data to image
-
-	// Transition image to be DST for copy operation
-	transitionImageLayout(logicalDevice, graphicsQueue, graphicsCommandPool,
-		texImage, VK_IMAGE_LAYOUT_UNDEFINED, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL);
-
-	// Copy image data
-	copyImageBuffer(logicalDevice, graphicsQueue, graphicsCommandPool, imageStagingBuffer, texImage, width, height);
-
-	// Transition image to be shader readable for shader usage
-	transitionImageLayout(logicalDevice, graphicsQueue, graphicsCommandPool,
-		texImage, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL);
-
-	// Add texture data to vector for reference
-	textureImages.push_back(texImage);
-	textureImageMemory.push_back(texImageMemory);
-
-	// Destroy staging buffer
-	vkDestroyBuffer(logicalDevice, imageStagingBuffer, nullptr);
-	vkFreeMemory(logicalDevice, imageStagingBufferMemory, nullptr);
-
-	// Return index to the new image
-	return textureImages.size() - 1;
-}
-
-VkImage MeshModel::createImage(uint32_t width, uint32_t height, VkFormat format, 
-	VkImageTiling tiling, VkImageUsageFlags usageFlags, VkMemoryPropertyFlags propFlags, 
-	VkDeviceMemory* imageMemory, VkPhysicalDevice& physicalDevice, VkDevice& logicalDevice)
-{
-	// Create image
-	// Image creation info
-	VkImageCreateInfo imageCreateInfo = {};
-	imageCreateInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
-	imageCreateInfo.imageType = VK_IMAGE_TYPE_2D; // Type of image (1D, 2D, 3D)
-	imageCreateInfo.extent.width = width; // w of image extent
-	imageCreateInfo.extent.height = height; // h of image extent
-	imageCreateInfo.extent.depth = 1; // Dep`th of image (just 1, no 3D aspect)
-	imageCreateInfo.mipLevels = 1; // Number of mipmap levels
-	imageCreateInfo.arrayLayers = 1; // Number of levels in image array
-	imageCreateInfo.format = format; // Format type of image
-	imageCreateInfo.tiling; // how image data should be tiled (arranged)
-	imageCreateInfo.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED; // Layout of image data on creation
-	imageCreateInfo.usage = usageFlags; // Bit flags defining what image will be used for
-	imageCreateInfo.samples = VK_SAMPLE_COUNT_1_BIT; // Number of samples for multi-sampling
-	imageCreateInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE; // Whether image can be shared between queues
-
-	VkImage image;
-	VkResult result = vkCreateImage(logicalDevice, &imageCreateInfo, nullptr, &image);
-	if (result != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to create an Image!");
-	}
-
-	// Create memory for image
-	// Get memory requirements for type of image
-	VkMemoryRequirements memoryRequirements;
-	vkGetImageMemoryRequirements(logicalDevice, image, &memoryRequirements);
-
-	// Allocate memory using image requirements and user defined properties
-	VkMemoryAllocateInfo memoryAllocInfo = {};
-	memoryAllocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-	memoryAllocInfo.allocationSize = memoryRequirements.size;
-	memoryAllocInfo.memoryTypeIndex = findMemoryTypeIndex(physicalDevice, memoryRequirements.memoryTypeBits, propFlags);
-
-	result = vkAllocateMemory(logicalDevice, &memoryAllocInfo, nullptr, imageMemory);
-	if (result != VK_SUCCESS)
-	{
-		throw std::runtime_error("Failed to allocate memory for image!");
-	}
-
-	// Connect memory to image
-	vkBindImageMemory(logicalDevice, image, *imageMemory, 0);
-
-	return image;
 }
